+++ src/ArubapecClient.php
@@ -9,6 +9,7 @@ use GuzzleHttp\ClientInterface;
 use Shellrent\Arubapec\Account\AccountClient;
 use Shellrent\Arubapec\AdditionalService\AdditionalServiceClient;
 use Shellrent\Arubapec\Auth\AuthClient;
+use Shellrent\Arubapec\Country\CountryClient;
 use Shellrent\Arubapec\Config\ClientConfig;
 
 final class ArubapecClient
@@ -22,6 +23,8 @@ final class ArubapecClient
     private readonly AccountClient $accountClient;
 
     private readonly AdditionalServiceClient $additionalServiceClient;
+
+    private readonly CountryClient $countryClient;
 
     public function __construct(
         ?ClientInterface $httpClient = null,
@@ -37,6 +40,7 @@ final class ArubapecClient
         $this->authClient = new AuthClient($this->httpClient);
         $this->accountClient = new AccountClient($this->httpClient);
         $this->additionalServiceClient = new AdditionalServiceClient($this->httpClient);
+        $this->countryClient = new CountryClient($this->httpClient);
     }
 
     public function auth(): AuthClient
@@ -52,6 +56,11 @@ final class ArubapecClient
     public function additionalService(): AdditionalServiceClient
     {
         return $this->additionalServiceClient;
+    }
+
+    public function country(): CountryClient
+    {
+        return $this->countryClient;
     }
 
     public function getHttpClient(): ClientInterface